# docs/api/api.yaml
openapi: 3.0.3
info:
  title: SmartMove API
  description: |
    REST + WebSocket API for SmartMove — a real estate analytics platform
    covering the Dubai and Egypt markets.

    ## Authentication
    All protected endpoints require a valid JWT access token delivered as an
    httpOnly cookie (`access_token`). Obtain tokens via `POST /api/auth/login/`.
    Tokens expire after **15 minutes**; use `POST /api/auth/refresh/` to renew.

    ## Markets
    Every data endpoint accepts an optional `market` query parameter:
    - `dubai`  — Dubai DLD transaction data
    - `egypt`  — Egypt property listings data
    - Omit to receive combined / cross-market results where applicable.

    ## Roles
    | Role    | Accessible endpoints |
    |---------|----------------------|
    | Admin   | All endpoints including `/monitoring/` |
    | Analyst | Dashboard, Upload, Reports, Predictions, Chatbot, Notifications |
    | Viewer  | Dashboard (read-only), Reports (read-only), Notifications |

  version: "1.0.0"
  contact:
    name: SmartMove Engineering
    url: https://github.com/SmartMove-Organization/SmartMove
servers:
  - url: http://localhost:8000/api
    description: Local development
  - url: https://smartmove.app/api
    description: Production (AKS)
...
security:
  - cookieAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token

  # ─── Reusable schemas ───────────────────────
  schemas:
...
    PaginatedMeta:
      type: object
      properties:
        count:
          type: integer
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
    Market:
      type: string
      enum: [dubai, egypt]

    # ── Auth ────────────────────────────────────
...
    ReportRequest:
      type: object
      required: [market, report_type]
      properties:
        market:
          $ref: "#/components/schemas/Market"
        report_type:
          type: string
          enum:
            [
              investment_opportunity,
              market_heat,
              investor_performance,
              market_risk,
              full_market,
            ]
        filters:
          type: object
          description: Same filter keys as DashboardFilters
          additionalProperties: true
        format:
          type: string
          enum: [pdf, excel]
          default: pdf
    ReportRecord:
      type: object
      properties:
        report_id:
          type: string
          format: uuid
        market:
          $ref: "#/components/schemas/Market"
        report_type:
          type: string
        format:
          type: string
          enum: [pdf, excel]
        status:
          type: string
          enum: [generating, ready, failed]
        download_url:
          type: string
          nullable: true
          description: Signed Azure Blob URL, valid for 1 hour
        created_at:
          type: string
          format: date-time

    # ── Predictions ────────────────────────────────
...
paths:
...
  /reports/:
    post:
      tags: [Reports]
      summary: Request a new report (async generation via Celery)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportRequest"
      responses:
        "202":
          description: Report generation queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportRecord"
    get:
      tags: [Reports]
      summary: List all reports for the current user
      parameters:
        - in: query
          name: market
          schema:
            $ref: "#/components/schemas/Market"
        - in: query
          name: status
          schema:
            type: string
            enum: [generating, ready, failed]
        - in: query
          name: page
          schema:
            type: integer
            default: 1
      responses:
        "200":
          description: Report list
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    $ref: "#/components/schemas/PaginatedMeta"
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/ReportRecord"
  /reports/{report_id}/:
    get:
      tags: [Reports]
      summary: Get status and download URL for a specific report
      parameters:
        - in: path
          name: report_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Report record with download_url when ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportRecord"
        "404":
          description: Report not found

  # ════════════════════════════════════════════
  # PREDICTIONS
  # ════════════════════════════════════════════
...